backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
Pillow>=10.0.0
geopy>=2.3.0
cryptography>=41.0.0
pycryptodome>=3.19.0
numpy>=1.24.0
opencv-python>=4.8.0
pytesseract>=0.3.10
pyzbar>=0.1.9
stegano>=0.9.9
exifread
requests>=2.31.0
python-magic
rarfile
py7zr
zstandard
lz4
python-whois
beautifulsoup4
dnspython
gunicorn>=21.2.0
